+++ webui/src/Buttons/index.jsx
@@ -4,7 +4,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { nanoid } from 'nanoid'
 import { InstancePresets } from './Presets'
 import { SocketContext, MyErrorBoundary, socketEmitPromise, UserConfigContext } from '../util'
-import { ButtonsGridPanel } from './ButtonGrid'
+import { ButtonsGridPanel } from './ButtonGridPanel'
 import { EditButton } from './EditButton'
 import { ActionRecorder } from './ActionRecorder'
 import { memo, useCallback, useContext, useRef, useState } from 'react'
